Forces use of GrOvalRenderer for ovals and rounded rectangles
This is always faster than GrAAConvexPathRenderer and GrDefaultPathRenderer fallbacks
provided that target is MSAA capable, regardless of fSampleCnt value provided by
description.

The real issue is that the value of fSampleCnt is not detected from actual target
capabilities, but is arbitrarily user-provided, which by default (0) leave to
sub-optmial rendering of oval shapes.